Add ability to export as xLights xmodel file.
This release adds the ability to export your star as an .xmodel file you
can import into xLights directly.
The star model in xLights is definitely more compact that the previous
method of exporting directly to your xLights show directory, at the expense
of the star models not perfectly matching the actual stars you generate.