Add support for showing pixel bodies.
When modeling layouts with square pixel nodes its helpful to be able to
see the body size.  This is just shown as a simple dashed line circle;
the pre-populated pixel size is the small dimension for a square node.
You can also use 0.911 for the larger size of those pixels.  Ideally you
can rotate pixels to make them fit.